Fix warnings and deprecations
* Rewrite platform.set to provides in class itself

Chef::Platform.set is deprecated from Chef 12.8.X and will become a hard error in Chef 13.

* Set default value for root_network_acl to []

The resource expects an Array, so the default value should represent this. Also, an each is performed on the value, so giving it an array it cleaner.

* Make enable_utf8 a True- or FalseClass type

It is used as a boolean value in the templates, so let's make it a boolean option instead of a string.